requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
pyyaml>=6.0
tqdm>=4.65.0
einops>=0.6.0
# MinkowskiEngine for sparse 3D convolutions (required for full functionality)
# Install separately: pip install MinkowskiEngine
# or: pip install git+https://github.com/NVIDIA/MinkowskiEngine.git
# MinkowskiEngine

# Optional: for faster Gaussian rendering
# diff-gaussian-rasterization (3DGS CUDA rasterizer)
# Install from: https://github.com/graphdeco-inria/gaussian-splatting

# Optional: Muon optimizer
# pip install git+https://github.com/KellerJordan/Muon.git
